+++ src-tauri/src/adapters/acp/mod.rs
@@ -1,5 +1,7 @@
 pub mod client;
+mod permission_flow;
 pub mod runner;
 pub mod session_update_mapper;
+mod terminal;
 pub mod transport;
 pub mod util;
